+++ Km_NodeGraphEN/constants.py
@@ -1,4 +1,4 @@
 """define constants"""
 
-# 
 BOOKMARK_KNOB_PREFIX = "KM_NGJ_V2_"
+KM_EN_FULL_NAME = "Km NodeGraph Easy Navigate v2.2"